Pivot direction and run comparisons through the full Core stack before checking matching
